use internal state and restructure
